server/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
alembic==1.14.0

